created trial module and jargon change
Class representing the trial data has been moved to a new module and a
distinction has been made between judged and unjudged trials. To better
emphasize the general nature of this code package, the language has
been modified to be more modality agnostic. Instead of referring to
trials as “displays”, they are now referred to as “similarity trials.”
The relevant classes, attributes, and variables have been updated to
reflect the new language.