Add FitRecord object.
The FitRecord object records and tracks the best fit restarts. The user can specify the number of top restarts to track. By default, only a single best restart is tracked. Nearly all of the tracking logic that was previously in the fit function has been moved into the new object.