Add argument to ProjectAttention.
Was getting "Could not find valid device for node." when computeing n_dim on the fly from shape of input. Work around is to add n_dim as a pre-computed attribute. Annoying, but shouldn't case any major problems since we should know the size of the embedding whenever we create this constraint.