Add approximate KL to variational method.
Add arguments and functionality to Variational layer. Arguments include the optional arguments kl_exact and kl_n_sample. If kl_exact=False, a sample-based method is used for computing KL divergence. This allows users to compute the KL divergence between distributions that do not have an analytical form.

In addition, it is now the responsiblity of Variational subclasses to implement a call method that samples the posterior and adds the KL loss to the graph. This change has been made to account for the fact that we cannot know all the relevant layer weight names ahead of time.